# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API for Nativos 3D pricing and production records
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /calculations:
    post:
      operationId: calculatePricing
      tags: [pricing]
      summary: Calculate production costs and channel prices
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalculationInput"
      responses:
        "200":
          description: Calculation, including any missing data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Calculation"
  /materials:
    get:
      operationId: listMaterials
      tags: [materials]
      summary: List saved materials
      responses:
        "200":
          description: Materials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Material"
    post:
      operationId: createMaterial
      tags: [materials]
      summary: Create a material
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialInput"
      responses:
        "201":
          description: Created material
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
...
  /quotes:
    get:
      operationId: listQuotes
      tags: [quotes]
      summary: List saved quotations
      responses:
        "200":
          description: Quotes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quote"
    post:
      operationId: createQuote
      tags: [quotes]
      summary: Save a quote snapshot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteInput"
      responses:
        "201":
          description: Created quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
components:
...
  schemas:
...
    Material:
      allOf:
        - $ref: "#/components/schemas/MaterialInput"
        - type: object
          properties:
            id: { type: integer }
            costPerGram: { type: number }
            createdAt: { type: string, format: date-time }
          required: [id, costPerGram, createdAt]
    MaterialInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        type: { type: string }
        brand: { type: string }
        color: { type: string }
        supplier: { type: string }
        netWeightGrams: { type: number, exclusiveMinimum: 0 }
        purchaseValue: { type: number, minimum: 0 }
        allocatedFreight: { type: number, minimum: 0 }
        stockGrams: { type: number, minimum: 0 }
        lowStockThreshold: { type: number, minimum: 0 }
        archived: { type: boolean }
      required: [name, type, brand, color, supplier, netWeightGrams, purchaseValue, allocatedFreight, stockGrams, lowStockThreshold, archived]
...
    Quote:
      type: object
      properties:
        id: { type: integer }
        customerName: { type: ["string", "null"] }
        productName: { type: string }
        quantity: { type: integer }
        unitPrice: { type: number }
        total: { type: number }
        status: { type: string, enum: [draft, sent, approved, rejected, expired] }
        validUntil: { type: ["string", "null"], format: date }
        notes: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, customerName, productName, quantity, unitPrice, total, status, validUntil, notes, createdAt]
    QuoteInput:
      type: object
      properties:
        customerName: { type: ["string", "null"] }
        productName: { type: string, minLength: 1 }
        quantity: { type: integer, minimum: 1 }
        unitPrice: { type: number, minimum: 0 }
        status: { type: string, enum: [draft, sent, approved, rejected, expired] }
        validUntil: { type: ["string", "null"], format: date }
        notes: { type: string }
      required: [customerName, productName, quantity, unitPrice, status, validUntil, notes]
    MaterialLine:
      type: object
      properties:
        materialId: { type: ["integer", "null"] }
        name: { type: string }
        costPerGram: { type: ["number", "null"] }
        usefulGrams: { type: number, minimum: 0 }
        supportsGrams: { type: number, minimum: 0 }
        brimGrams: { type: number, minimum: 0 }
        purgeGrams: { type: number, minimum: 0 }
        wasteMode: { type: string, enum: [none, grams, percent] }
        wasteValue: { type: number, minimum: 0 }
        slicerTotalGrams: { type: ["number", "null"] }
      required: [materialId, name, costPerGram, usefulGrams, supportsGrams, brimGrams, purgeGrams, wasteMode, wasteValue, slicerTotalGrams]
    CalculationInput:
      type: object
      properties:
        name: { type: string }
        quantity: { type: integer, minimum: 1 }
        piecesPerRun: { type: integer, minimum: 1 }
        materials:
          type: array
          items: { $ref: "#/components/schemas/MaterialLine" }
        printMinutes: { type: ["number", "null"], minimum: 0 }
        setupMinutesPerRun: { type: number, minimum: 0 }
        finishingMinutesPerUnit: { type: number, minimum: 0 }
        jobPrepMinutes: { type: number, minimum: 0 }
        printerId: { type: ["integer", "null"] }
        averagePowerWatts: { type: ["number", "null"], minimum: 0 }
        measuredKwh: { type: ["number", "null"], minimum: 0 }
        energyRate: { type: ["number", "null"], minimum: 0 }
        machinePurchaseValue: { type: ["number", "null"], minimum: 0 }
        machineResidualValue: { type: ["number", "null"], minimum: 0 }
        machineUsefulLifeHours: { type: ["number", "null"], minimum: 0 }
        machineMaintenancePerHour: { type: ["number", "null"], minimum: 0 }
        machineEnabled: { type: boolean }
        hourlyLaborRate: { type: ["number", "null"], minimum: 0 }
        monthlyFixedExpenses: { type: ["number", "null"], minimum: 0 }
        productiveHoursMonthly: { type: ["number", "null"], minimum: 0 }
        failureRate: { type: ["number", "null"], minimum: 0, maximum: 0.99 }
        packagePerUnit: { type: number, minimum: 0 }
        packagePerOrder: { type: number, minimum: 0 }
        componentsPerUnit: { type: number, minimum: 0 }
        licensePerUnit: { type: number, minimum: 0 }
        licensePerOrder: { type: number, minimum: 0 }
        sellerShipping: { type: number, minimum: 0 }
        otherPerUnit: { type: number, minimum: 0 }
        otherPerOrder: { type: number, minimum: 0 }
        percentFees: { type: number, minimum: 0, maximum: 0.99 }
        fixedFeePerOrder: { type: number, minimum: 0 }
        fixedFeePerUnit: { type: number, minimum: 0 }
        desiredMargin: { type: number, minimum: 0, maximum: 0.99 }
        rounding: { type: string, enum: [cent, half, whole, ending90, ending99] }
        practicedUnitPrice: { type: ["number", "null"], minimum: 0 }
      required: [name, quantity, piecesPerRun, materials, printMinutes, setupMinutesPerRun, finishingMinutesPerUnit, jobPrepMinutes, printerId, averagePowerWatts, measuredKwh, energyRate, machinePurchaseValue, machineResidualValue, machineUsefulLifeHours, machineMaintenancePerHour, machineEnabled, hourlyLaborRate, monthlyFixedExpenses, productiveHoursMonthly, failureRate, packagePerUnit, packagePerOrder, componentsPerUnit, licensePerUnit, licensePerOrder, sellerShipping, otherPerUnit, otherPerOrder, percentFees, fixedFeePerOrder, fixedFeePerUnit, desiredMargin, rounding, practicedUnitPrice]
    CostLine:
      type: object
      properties:
        label: { type: string }
        amount: { type: number }
      required: [label, amount]
    ChannelPrice:
      type: object
      properties:
        channel: { type: string }
        breakEvenPerUnit: { type: ["number", "null"] }
        recommendedPerUnit: { type: ["number", "null"] }
        total: { type: ["number", "null"] }
        estimatedResult: { type: ["number", "null"] }
        estimatedMargin: { type: ["number", "null"] }
      required: [channel, breakEvenPerUnit, recommendedPerUnit, total, estimatedResult, estimatedMargin]
    Calculation:
      type: object
      properties:
        complete: { type: boolean }
        missing: { type: array, items: { type: string } }
        errors: { type: array, items: { type: string } }
        quantity: { type: integer }
        runs: { type: integer }
        manufacturingCost: { type: ["number", "null"] }
        completeCost: { type: ["number", "null"] }
        unitCost: { type: ["number", "null"] }
        materialGrams: { type: ["number", "null"] }
        costs: { type: array, items: { $ref: "#/components/schemas/CostLine" } }
        channels: { type: array, items: { $ref: "#/components/schemas/ChannelPrice" } }
        legacyComparison:
          type: object
          properties:
            cost: { type: ["number", "null"] }
            legacyMinimum: { type: ["number", "null"] }
            legacyWholesale: { type: ["number", "null"] }
            legacyMarketplace: { type: ["number", "null"] }
          required: [cost, legacyMinimum, legacyWholesale, legacyMarketplace]
        assumptions: { type: array, items: { type: string } }
      required: [complete, missing, errors, quantity, runs, manufacturingCost, completeCost, unitCost, materialGrams, costs, channels, legacyComparison, assumptions]