# lib/api-zod/src/generated/types/calculationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Nativos 3D pricing and production records
 * OpenAPI spec version: 0.1.0
 */
import type { CalculationInputRounding } from './calculationInputRounding';
import type { MaterialLine } from './materialLine';

export interface CalculationInput {
  name: string;
  /** @minimum 1 */
  quantity: number;
  /** @minimum 1 */
  piecesPerRun: number;
  materials: MaterialLine[];
  /**
     * @minimum 0
     * @nullable
     */
  printMinutes: number | null;
  /** @minimum 0 */
  setupMinutesPerRun: number;
  /** @minimum 0 */
  finishingMinutesPerUnit: number;
  /** @minimum 0 */
  jobPrepMinutes: number;
  /** @nullable */
  printerId: number | null;
  /**
     * @minimum 0
     * @nullable
     */
  averagePowerWatts: number | null;
  /**
     * @minimum 0
     * @nullable
     */
  measuredKwh: number | null;
  /**
     * @minimum 0
     * @nullable
     */
  energyRate: number | null;
  /**
     * @minimum 0
     * @nullable
     */
  machinePurchaseValue: number | null;
  /**
     * @minimum 0
     * @nullable
     */
  machineResidualValue: number | null;
  /**
     * @minimum 0
     * @nullable
     */
  machineUsefulLifeHours: number | null;
  /**
     * @minimum 0
     * @nullable
     */
  machineMaintenancePerHour: number | null;
  machineEnabled: boolean;
  /**
     * @minimum 0
     * @nullable
     */
  hourlyLaborRate: number | null;
  /**
     * @minimum 0
     * @nullable
     */
  monthlyFixedExpenses: number | null;
  /**
     * @minimum 0
     * @nullable
     */
  productiveHoursMonthly: number | null;
  /**
     * @minimum 0
     * @maximum 0.99
     * @nullable
     */
  failureRate: number | null;
  /** @minimum 0 */
  packagePerUnit: number;
  /** @minimum 0 */
  packagePerOrder: number;
  /** @minimum 0 */
  componentsPerUnit: number;
  /** @minimum 0 */
  licensePerUnit: number;
  /** @minimum 0 */
  licensePerOrder: number;
  /** @minimum 0 */
  sellerShipping: number;
  /** @minimum 0 */
  otherPerUnit: number;
  /** @minimum 0 */
  otherPerOrder: number;
  /**
     * @minimum 0
     * @maximum 0.99
     */
  percentFees: number;
  /** @minimum 0 */
  fixedFeePerOrder: number;
  /** @minimum 0 */
  fixedFeePerUnit: number;
  /**
     * @minimum 0
     * @maximum 0.99
     */
  desiredMargin: number;
  rounding: CalculationInputRounding;
  /**
     * @minimum 0
     * @nullable
     */
  practicedUnitPrice: number | null;
}