# lib/api-zod/src/generated/types/costLine.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Nativos 3D pricing and production records
 * OpenAPI spec version: 0.1.0
 */

export interface CostLine {
  label: string;
  amount: number;
}