# lib/api-zod/src/generated/types/materialInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Nativos 3D pricing and production records
 * OpenAPI spec version: 0.1.0
 */

export interface MaterialInput {
  /** @minLength 1 */
  name: string;
  type: string;
  brand: string;
  color: string;
  supplier: string;
  /** @exclusiveMinimum 0 */
  netWeightGrams: number;
  /** @minimum 0 */
  purchaseValue: number;
  /** @minimum 0 */
  allocatedFreight: number;
  /** @minimum 0 */
  stockGrams: number;
  /** @minimum 0 */
  lowStockThreshold: number;
  archived: boolean;
}